updates home structure, adds about page
- includes edits to about content
- content that was at `/` is now on `/about`
- home embeds content of `/about`
