+++ src/lib/components/AboutPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AboutPage.svelte'
