+++ src/lib/components/Header/Header.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+</script>
+
+<div class="header">
+  <a href="/"><img src="/wbd-logo.svg" alt="whatbrentdo logo" /></a>
+</div>
+
+<style>
+  .header {
+    width: 3rem;
+    height: 3rem;
+    margin: 0 auto 2rem;
+  }
+</style>
+
