+++ src/lib/components/RevealText/RevealText.svelte
