+++ src/lib/components/RevealText/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RevealText.svelte'
