--- src/routes/Header.svelte
@@ -1,13 +1,0 @@
-<div class='header'>
-  <h1>whatbrentdo</h1>
-</div>
-
-<style>
-  .header {
-    background: url('/wbd-logo.svg') no-repeat;
-    height: 3rem;
-    width: 3rem;
-    margin: 0 auto 2rem;
-    text-indent: -9999px;
-  }
-</style>
